Tiffany's updates
style.css was updated for aesthetics: changed calculator background color, margin, and border radius. Fixed the font-sizes, font-weight, and margins of certain areas.
tip_calc.js was updated and the htmlString was split into two to avoid shifting of calculator to left (but now calculator shifts up a little bit the first time you use it)
tip_calc.html was updated to match the javascript edit so that htmlString1 and htmlString2 has target locations.